requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
biopython==1.77
celery==4.4.7
certifi==2020.6.20
cycler==0.10.0
Django==3.1
django-debug-toolbar==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
et-xmlfile==1.0.1
inflection==0.5.0
jdcal==1.4.1
joblib==0.16.0
kiwisolver==1.2.0
kombu==4.6.11
matplotlib==3.3.1
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.48.2
vine==1.3.0
xlrd==1.2.0
